wip start extracting resource field into a shared component
